Page widget with BTR changes
