+++ now-v2.json
@@ -1,0 +1,11 @@
+{
+	"version": 2,
+	"name": "dojo.io",
+	"builds": [
+		{
+			"src": "package.json",
+			"use": "@now/static-build",
+			"config": { "useBuildUtils": "@now/build-utils@canary" }
+		}
+	]
+}
