+++ now.json
@@ -1,11 +1,4 @@
 {
-  "version": 2,
-  "name": "dojo.io",
-  "builds": [
-    {
-			"src": "package.json",
-			"use": "@now/static-build",
-			"config": { "useBuildUtils": "@now/build-utils@canary" }
-		}
-	]
+	"version": 1,
+	"type": "static"
 }
