+++ package.json
@@ -2,40 +2,34 @@
   "name": "dojo.io",
   "version": "1.0.0",
   "scripts": {
-    "build:content": "node pipeline",
-    "build:site": "dojo build",
-    "build": "run-s build:content build:site",
-    "dev:content": "chokidar \"content/**/*.*\" -c \"npm run build:content {event} {path}\"",
-    "dev:site": "dojo build -m dev -w -s",
-    "dev": "npm-run-all build:content --parallel dev:**",
+    "build": "dojo build",
+    "dev": "dojo build -m dev -w",
     "intern": "intern",
     "intern:ci": "intern config=@ci",
     "lint:tslint": "tslint -p .",
     "lint:prettier": "prettier -l \"{src,tests,scripts}/**/*.{ts,tsx}\"",
     "lint": "run-p lint:*",
-    "now-build": "npm run build && mv output/dist dist",
+    "now-build": "npm run build && mv ./output/dist /public",
     "prettier": "prettier --write \"{src,tests,scripts}/**/*.{ts,tsx}\"",
     "test:build": "tsc && dojo build -m unit -l",
-    "test:build:static": "cpx \"{src,tests,content}/**/*.{html,png,jpg,gif,json,xml,css,txt,d.ts,svg,ico}\" _build",
     "test:clean": "shx rm -Rf _build",
     "test:ci": "run-s test:build test:build:static",
     "test": "run-s test:build test:build:static"
   },
   "dependencies": {
-    "@dojo/framework": "file:deps/dojo-framework-4.0.5.tgz",
-    "tslib": "~1.8.1"
+    "@dojo/framework": "next",
+    "tslib": "1.9.3"
   },
   "devDependencies": {
     "@dojo/cli": "^4.0.0",
-    "@dojo/cli-build-app": "file:deps/dojo-cli-build-app-4.0.5.tgz",
-    "@dojo/cli-test-intern": "^4.0.0",
+    "@dojo/cli-build-app": "next",
     "@dojo/scripts": "^3.1.0",
-    "@dojo/webpack-contrib": "file:deps/dojo-webpack-contrib-4.0.5.tgz",
+    "@dojo/webpack-contrib": "next",
+    "@mapbox/rehype-prism": "^0.3.0",
     "@types/fs-extra": "^5.0.4",
     "@types/mockery": "^1.4.29",
     "@types/node": "~9.6.5",
     "@types/sinon": "~4.3.3",
-    "@mapbox/rehype-prism": "^0.3.0",
     "chokidar-cli": "^1.2.1",
     "codecov": "^3.0.4",
     "cpx": "^1.5.0",
