--- scripts/logger.ts
@@ -1,3 +1,0 @@
-export function info(...args: any[]) {
-	console.info(...args);
-}
