--- scripts/tsconfig.json
@@ -1,30 +1,0 @@
-{
-	"compilerOptions": {
-		"declaration": false,
-		"experimentalDecorators": true,
-		"noUnusedParameters": true,
-		"lib": [ "es2015", "dom", "es2016.array.include" ],
-		"module": "commonjs",
-		"moduleResolution": "node",
-		"outDir": "_build/tools",
-		"strictNullChecks": true,
-		"strictFunctionTypes": true,
-		"noUnusedLocals": true,
-		"noImplicitThis": true,
-		"noEmitOnError": true,
-		"noImplicitAny": true,
-		"removeComments": false,
-		"importHelpers": true,
-		"sourceMap": true,
-		"downlevelIteration": true,
-		"strict": true,
-		"target": "es5",
-		"types": [
-			"node"
-		],
-		"baseUrl": "."
-	},
-	"files": [
-		"compile.ts"
-	]
-}
