+++ src/App.tsx
@@ -3,9 +3,9 @@ import { tsx } from '@dojo/framework/widget-core/tsx';
 import Outlet from '@dojo/framework/routing/Outlet';
 
 import Menu from './widgets/Menu';
+import Page from './widgets/Page';
 import Home from './pages/Home';
 import Blog from './pages/Blog';
-import Documentation from './pages/Documentation';
 import Examples from './pages/Examples';
 import Playground from './pages/Playground';
 import Community from './pages/Community';
@@ -20,10 +20,19 @@ export default class App extends WidgetBase {
 				<div classes={[css.content]}>
 					<Outlet key="home" id="home" renderer={() => <Home />} />
 					<Outlet key="blog" id="blog" renderer={() => <Blog />} />
-					<Outlet key="documentation" id="documentation" renderer={() => <Documentation />} />
 					<Outlet key="examples" id="examples" renderer={() => <Examples />} />
 					<Outlet key="playground" id="playground" renderer={() => <Playground />} />
 					<Outlet key="community" id="community" renderer={() => <Community />} />
+					<Outlet
+						key="page"
+						id="page"
+						renderer={(matchDetails) => {
+							if (matchDetails.isExact) {
+								const path = `${matchDetails.params.section}/${matchDetails.params.page}`;
+								return <Page path={path} />;
+							}
+						}}
+					/>
 				</div>
 			</div>
 		);
