+++ src/index.html
@@ -6,6 +6,7 @@
 	<meta name="viewport" content="width=device-width, initial-scale=1">
 	<meta name="description" content="The home of dojo. Documentation. Examples. Community.">
 	<link rel="icon" href="./assets/favicon.ico" />
+	<link rel="stylesheet" type="text/css" href="./assets/prism.css" />
 </head>
 <body>
 	<div id="root"></div>
