+++ src/main.tsx
@@ -2,6 +2,7 @@ import renderer from '@dojo/framework/widget-core/vdom';
 import Registry from '@dojo/framework/widget-core/Registry';
 import { tsx } from '@dojo/framework/widget-core/tsx';
 import { registerRouterInjector } from '@dojo/framework/routing/RouterInjector';
+import { StateHistory } from '@dojo/framework/routing/history/StateHistory';
 
 import Aside from './widgets/content/Aside';
 import CodeBlock from './widgets/code/CodeBlock';
@@ -15,7 +16,7 @@ import App from './App';
 
 const registry = new Registry();
 
-registerRouterInjector(routes, registry);
+registerRouterInjector(routes, registry, { HistoryManager: StateHistory });
 
 registry.define('docs-aside', Aside);
 registry.define('docs-codeblock', CodeBlock);
