+++ src/scripts/regions/region-matchers/inline-hash.ts
