+++ tsconfig.json
@@ -26,7 +26,6 @@
 		"./src/**/*.ts",
 		"./src/**/*.tsx",
 		"./tests/**/*.ts",
-		"./tests/**/*.tsx",
-		"./scripts/**/*.ts"
+		"./tests/**/*.tsx"
 	]
 }
